requirements.txt
dj-database-url==0.4.1
django==1.11.1
django-crispy-forms==1.6.1
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2